test linux
git-svn-id: svn+ssh://oberon/svn/OMiSCID-Dev@908 c4bd1236-c4f7-0310-a2fe-a60683d9aa4a

+++ System/Portage.cpp
@@ -78,7 +78,7 @@ int Omiscid::gettimeofday(struct timeval * tv,struct timezone * tz )
 	struct _timeb tb;
 	_ftime(&tb);
 
-	tv->tv_sec  = tb.time;
+	tv->tv_sec  = (long)tb.time;
 	tv->tv_usec = tb.millitm * (long)1000;
 
 	return 0;
